fastapi-msc-test/app/requirements.txt
# Web & Server
fastapi
uvicorn
httpx[http2]
python-json-logger

## OpenTelemetry Core
opentelemetry-distro
opentelemetry-exporter-otlp

